fix: local compilation errors - Double/Float types, missing imports, Composable annotations
- Fix BoardView.kt: use Float literals (0.07f) instead of Double
- Fix BoardView.kt: use .sp extension instead of TextUnit constructor
- Fix AnalysisPanels.kt: use .toFloat() for Path methods
- Fix LizzieMobileApp.kt: add dp import
- Fix SettingsScreen.kt: add @Composable to helper functions
- Add commonMain source set with Compose dependencies

+++ lizzie-mobile/composeApp/src/androidMain/kotlin/com/lizzie/android/ui/analysis/AnalysisPanels.kt
@@ -50,8 +50,8 @@ fun WinrateGraph(
         dataPoints.forEachIndexed { i, winrate ->
             val x = i * stepX
             // winrate: 0-100, map to y (0 at bottom = 100%, height at top = 0%)
-            val y = graphHeight - ((winrate / 100f) * (graphHeight - 2 * padding)) - padding
-            if (i == 0) path.moveTo(x, y) else path.lineTo(x, y)
+            val y = graphHeight - ((winrate / 100.0) * (graphHeight - 2.0 * padding)) - padding
+            if (i == 0) path.moveTo(x, y.toFloat()) else path.lineTo(x, y.toFloat())
         }
 
         drawPath(
